refactor: extract CompletionProvider(s) to their own files
