+++ src/main/kotlin/com/github/lppedd/cc/completion/providers/FooterTypeCompletionProvider.kt
@@ -5,10 +5,7 @@ package com.github.lppedd.cc.completion.providers
 import com.github.lppedd.cc.MAX_ITEMS_PER_PROVIDER
 import com.github.lppedd.cc.api.CommitFooterProvider
 import com.github.lppedd.cc.api.FOOTER_EP
-import com.github.lppedd.cc.api.ProviderPresentation
-import com.github.lppedd.cc.completion.Priority
 import com.github.lppedd.cc.completion.resultset.ResultSet
-import com.github.lppedd.cc.configuration.CCConfigService
 import com.github.lppedd.cc.lookupElement.CommitFooterTypeLookupElement
 import com.github.lppedd.cc.parser.FooterContext.FooterTypeContext
 import com.github.lppedd.cc.psiElement.CommitFooterTypePsiElement
@@ -23,7 +20,7 @@ import org.jetbrains.annotations.ApiStatus
 internal class FooterTypeCompletionProvider(
     private val project: Project,
     private val context: FooterTypeContext,
-) : CommitCompletionProvider<CommitFooterProvider> {
+) : CompletionProvider<CommitFooterProvider> {
   override val providers: List<CommitFooterProvider> = FOOTER_EP.getExtensions(project)
   override val stopHere = false
 
@@ -44,20 +41,4 @@ internal class FooterTypeCompletionProvider(
       .distinctBy(CommitFooterTypeLookupElement::getLookupString)
       .forEach(rs::addElement)
   }
-}
-
-internal class FooterTypeProviderWrapper(
-    project: Project,
-    private val provider: CommitFooterProvider,
-) : ProviderWrapper {
-  private val config = CCConfigService.getInstance(project)
-
-  override fun getId(): String =
-    provider.getId()
-
-  override fun getPresentation(): ProviderPresentation =
-    provider.getPresentation()
-
-  override fun getPriority() =
-    Priority(config.getProviderOrder(provider))
 }
