+++ src/main/kotlin/com/github/lppedd/cc/completion/providers/ScopeCompletionProvider.kt
@@ -4,11 +4,8 @@ package com.github.lppedd.cc.completion.providers
 
 import com.github.lppedd.cc.MAX_ITEMS_PER_PROVIDER
 import com.github.lppedd.cc.api.CommitScopeProvider
-import com.github.lppedd.cc.api.ProviderPresentation
 import com.github.lppedd.cc.api.SCOPE_EP
-import com.github.lppedd.cc.completion.Priority
 import com.github.lppedd.cc.completion.resultset.ResultSet
-import com.github.lppedd.cc.configuration.CCConfigService
 import com.github.lppedd.cc.lookupElement.CommitScopeLookupElement
 import com.github.lppedd.cc.parser.CommitContext.ScopeCommitContext
 import com.github.lppedd.cc.psiElement.CommitScopePsiElement
@@ -23,7 +20,7 @@ import org.jetbrains.annotations.ApiStatus
 internal class ScopeCompletionProvider(
     private val project: Project,
     private val context: ScopeCommitContext,
-) : CommitCompletionProvider<CommitScopeProvider> {
+) : CompletionProvider<CommitScopeProvider> {
   override val providers: List<CommitScopeProvider> = SCOPE_EP.getExtensions(project)
   override val stopHere = false
 
@@ -44,20 +41,4 @@ internal class ScopeCompletionProvider(
       .distinctBy(CommitScopeLookupElement::getLookupString)
       .forEach(rs::addElement)
   }
-}
-
-internal class ScopeProviderWrapper(
-    project: Project,
-    private val provider: CommitScopeProvider,
-) : ProviderWrapper {
-  private val config = CCConfigService.getInstance(project)
-
-  override fun getId(): String =
-    provider.getId()
-
-  override fun getPresentation(): ProviderPresentation =
-    provider.getPresentation()
-
-  override fun getPriority() =
-    Priority(config.getProviderOrder(provider))
 }
